rename package io.sbk.api -> io.sbk.ram in SBK-RAM module.
