+++ sbk-ram/src/main/java/io/sbk/ram/RamParameters.java
@@ -8,7 +8,9 @@
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.sbk.api;
+package io.sbk.ram;
+
+import io.sbk.api.Action;
 
 public interface RamParameters {
