+++ sbk-ram/src/main/java/io/sbk/ram/impl/LatenciesRecordsBenchmark.java
@@ -8,7 +8,7 @@
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.sbk.api.impl;
+package io.sbk.ram.impl;
 
 import io.sbk.api.Benchmark;
 import io.sbk.api.RWCount;
